pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "le-shop"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1",
    "python-decouple",
    "djangorestframework",    
    "django-filter",
    "django-allauth[mfa]",
    "dj-database-url",
    "django-csp",    
    "cloudinary",
    "django-cloudinary-storage",
    "psycopg",
    "gunicorn",
    "django-unfold",
    "PyJWT>=2.8",
    "whitenoise",

]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "types-requests",
    "pytest",
    "pytest-django",
    "factory-boy",
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "mypy",
    "django-stubs",
    "djangorestframework-stubs",
    "types-requests",
    "pytest",
    "pytest-django",
    "factory-boy",
]

[tool.ruff]
exclude = [".git", ".venv", "migrations", "node_modules", "src/static"]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = []

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = false
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "le_shop.settings"

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["src"]
include = ["apps*", "le_shop*"]
